devices/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.3.0            # via django
certifi==2020.11.8        # via -r requirements.in, fvhiot
click==7.1.2              # via flask
django==3.1.3             # via -r requirements.in, djangorestframework, drf-nested-routers
djangorestframework==3.12.2  # via -r requirements.in, drf-nested-routers
drf-nested-routers==0.92.1  # via -r requirements.in
flask==1.1.2              # via fvhiot
git+https://github.com/ForumViriumHelsinki/FVHIoT-python.git#egg=fvhiot  # via -r requirements.in
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
kafka-python==2.0.2       # via fvhiot
markupsafe==1.1.1         # via jinja2
msgpack==1.0.0            # via fvhiot
psycopg2-binary==2.8.6    # via -r requirements.in
pytz==2020.4              # via django
sqlparse==0.4.1           # via django
werkzeug==1.0.1           # via flask
